Corrigido erro da validação de compra

+++ LojaAPI/LojaAPI/Models/Misc/Detalhes.cs
@@ -8,12 +8,12 @@ namespace LojaAPI.Models.Misc
     public class Detalhes
     {
         public Detalhes() { }
-        public Detalhes(float valor, String estado)
+        public Detalhes(double valor, String estado)
         {
             Valor = valor;
             Estado = estado;
         }
-        public float Valor { get; set; }
+        public double Valor { get; set; }
         public String Estado { get; set; }
     }
 }
